Further refactored config handling to better support multiple subcomands
