+++ wordfence/cli/vulnscan/vulnscan.py
@@ -1,2 +1,11 @@
-def main(config) -> int:
-    return 0
+from ..subcommands import Subcommand
+
+
+class VulnScanSubcommand(Subcommand):
+
+    def invoke(self) -> int:
+        print('Test')
+        return 0
+
+
+factory = VulnScanSubcommand
